+++ com.unity.render-pipelines.universal/Editor/ShaderGUI/BaseShaderGUI.cs
@@ -293,11 +293,11 @@ namespace UnityEditor
             EditorGUI.BeginDisabledGroup(isDisabled);
             EditorGUI.indentLevel += indentLevel;
             EditorGUI.BeginChangeCheck();
-            EditorGUI.showMixedValue = prop.hasMixedValue;
+            MaterialEditor.BeginProperty(prop);
             bool newValue = EditorGUILayout.Toggle(styles, prop.floatValue == 1);
             if (EditorGUI.EndChangeCheck())
                 prop.floatValue = newValue ? 1.0f : 0.0f;
-            EditorGUI.showMixedValue = false;
+            MaterialEditor.EndProperty();
             EditorGUI.indentLevel -= indentLevel;
             EditorGUI.EndDisabledGroup();
         }
@@ -314,7 +314,7 @@ namespace UnityEditor
                     BlendMode blendMode = (BlendMode)material.GetFloat(Property.BlendMode);
                     var isDisabled = blendMode == BlendMode.Multiply || blendMode == BlendMode.Premultiply;
                     if (!isDisabled)
-                        DrawFloatToggleProperty(Styles.preserveSpecularText, preserveSpecProp, 2, isDisabled);
+                        DrawFloatToggleProperty(Styles.preserveSpecularText, preserveSpecProp, 1, isDisabled);
                 }
             }
             DoPopup(Styles.cullingText, cullingProp, Styles.renderFaceNames);
@@ -776,6 +776,9 @@ namespace UnityEditor
         {
             const int kInterFieldPadding = 2;
 
+            MaterialEditor.BeginProperty(prop1);
+            MaterialEditor.BeginProperty(prop2);
+
             Rect rect = EditorGUILayout.GetControlRect();
             EditorGUI.PrefixLabel(rect, title);
 
@@ -804,6 +807,9 @@ namespace UnityEditor
             EditorGUIUtility.labelWidth = preLabelWidth;
 
             EditorGUI.showMixedValue = false;
+
+            MaterialEditor.EndProperty();
+            MaterialEditor.EndProperty();
         }
 
         public void DoPopup(GUIContent label, MaterialProperty property, string[] options)
@@ -815,6 +821,10 @@ namespace UnityEditor
         // Helper to show texture and color properties
         public static Rect TextureColorProps(MaterialEditor materialEditor, GUIContent label, MaterialProperty textureProp, MaterialProperty colorProp, bool hdr = false)
         {
+            MaterialEditor.BeginProperty(textureProp);
+            if (colorProp != null)
+                MaterialEditor.BeginProperty(colorProp);
+
             Rect rect = EditorGUILayout.GetControlRect();
             EditorGUI.showMixedValue = textureProp.hasMixedValue;
             materialEditor.TexturePropertyMiniThumbnail(rect, textureProp, label.text, label.tooltip);
@@ -838,6 +848,10 @@ namespace UnityEditor
                 }
                 EditorGUI.showMixedValue = false;
             }
+
+            if (colorProp != null)
+                MaterialEditor.EndProperty();
+            MaterialEditor.EndProperty();
 
             return rect;
         }